properly move config errors into the dbt exception hierarchy
