+++ dbt/task/debug.py
@@ -3,6 +3,7 @@ import pprint
 from dbt.logger import GLOBAL_LOGGER as logger
 import dbt.clients.system
 import dbt.config
+import dbt.exceptions
 
 from dbt.task.base_task import BaseTask
 
@@ -30,13 +31,13 @@ class DebugTask(BaseTask):
         try:
             project = dbt.config.Project.from_current_directory()
             project_profile = project.profile_name
-        except dbt.config.DbtConfigError as exc:
+        except dbt.exceptions.DbtConfigError as exc:
             project = 'ERROR loading project: {!s}'.format(exc)
 
         # log the profile we decided on as well, if it's available.
         try:
             profile = dbt.config.Profile.from_args(self.args, project_profile)
-        except dbt.config.DbtConfigError as exc:
+        except dbt.exceptions.DbtConfigError as exc:
             profile = 'ERROR loading profile: {!s}'.format(exc)
 
         logger.info("args: {}".format(self.args))
