+++ test/unit/test_config.py
@@ -10,6 +10,7 @@ import mock
 import yaml
 
 import dbt.config
+import dbt.exceptions
 from dbt.contracts.connection import PostgresCredentials, RedshiftCredentials
 from dbt.contracts.project import PackageConfig
 
@@ -250,7 +251,7 @@ class TestProfile(BaseConfigTest):
 
     def test_missing_type(self):
         del self.default_profile_data['default']['outputs']['postgres']['type']
-        with self.assertRaises(dbt.config.DbtProfileError) as exc:
+        with self.assertRaises(dbt.exceptions.DbtProfileError) as exc:
             profile = dbt.config.Profile.from_raw_profiles(
                 self.default_profile_data, 'default'
             )
@@ -260,7 +261,7 @@ class TestProfile(BaseConfigTest):
 
     def test_bad_type(self):
         self.default_profile_data['default']['outputs']['postgres']['type'] = 'invalid'
-        with self.assertRaises(dbt.config.DbtProfileError) as exc:
+        with self.assertRaises(dbt.exceptions.DbtProfileError) as exc:
             profile = dbt.config.Profile.from_raw_profiles(
                 self.default_profile_data, 'default'
             )
@@ -270,7 +271,7 @@ class TestProfile(BaseConfigTest):
 
     def test_invalid_credentials(self):
         del self.default_profile_data['default']['outputs']['postgres']['host']
-        with self.assertRaises(dbt.config.DbtProfileError) as exc:
+        with self.assertRaises(dbt.exceptions.DbtProfileError) as exc:
             profile = dbt.config.Profile.from_raw_profiles(
                 self.default_profile_data, 'default'
             )
@@ -280,7 +281,7 @@ class TestProfile(BaseConfigTest):
 
     def test_target_missing(self):
         del self.default_profile_data['default']['target']
-        with self.assertRaises(dbt.config.DbtProfileError) as exc:
+        with self.assertRaises(dbt.exceptions.DbtProfileError) as exc:
             profile = dbt.config.Profile.from_raw_profiles(
                 self.default_profile_data, 'default'
             )
@@ -288,7 +289,7 @@ class TestProfile(BaseConfigTest):
         self.assertIn('default', str(exc.exception))
 
     def test_profile_invalid_project(self):
-        with self.assertRaises(dbt.config.DbtProjectError) as exc:
+        with self.assertRaises(dbt.exceptions.DbtProjectError) as exc:
             profile = dbt.config.Profile.from_raw_profiles(
                 self.default_profile_data, 'invalid-profile'
             )
@@ -298,7 +299,7 @@ class TestProfile(BaseConfigTest):
         self.assertIn('invalid-profile', str(exc.exception))
 
     def test_profile_invalid_target(self):
-        with self.assertRaises(dbt.config.DbtProfileError) as exc:
+        with self.assertRaises(dbt.exceptions.DbtProfileError) as exc:
             profile = dbt.config.Profile.from_raw_profiles(
                 self.default_profile_data, 'default', target_override='nope',
             )
@@ -309,7 +310,7 @@ class TestProfile(BaseConfigTest):
         self.assertIn('- with-vars', str(exc.exception))
 
     def test_no_outputs(self):
-        with self.assertRaises(dbt.config.DbtProfileError) as exc:
+        with self.assertRaises(dbt.exceptions.DbtProfileError) as exc:
             profile = dbt.config.Profile.from_raw_profiles(
                 {'some-profile': {'target': 'blah'}}, 'some-profile'
             )
@@ -335,7 +336,7 @@ class TestProfile(BaseConfigTest):
     def test_invalid_env_vars(self):
         self.env_override['env_value_port'] = 'hello'
         with mock.patch.dict(os.environ, self.env_override):
-            with self.assertRaises(dbt.config.DbtProfileError) as exc:
+            with self.assertRaises(dbt.exceptions.DbtProfileError) as exc:
                 dbt.config.Profile.from_raw_profile_info(
                     self.default_profile_data['default'],
                     'default',
@@ -443,7 +444,7 @@ class TestProfileFile(BaseFileTest):
         self.assertEqual(profile, from_raw)
 
     def test_no_profile(self):
-        with self.assertRaises(dbt.config.DbtProjectError) as exc:
+        with self.assertRaises(dbt.exceptions.DbtProjectError) as exc:
             dbt.config.Profile.from_args(self.args)
         self.assertIn('no profile was specified', str(exc.exception))
 
@@ -661,14 +662,14 @@ class TestProject(BaseConfigTest):
 
     def test_invalid_project_name(self):
         self.default_project_data['name'] = 'invalid-project-name'
-        with self.assertRaises(dbt.config.DbtProjectError) as exc:
+        with self.assertRaises(dbt.exceptions.DbtProjectError) as exc:
             project = dbt.config.Project.from_project_config(
                 self.default_project_data
             )
         self.assertIn('invalid-project-name', str(exc.exception))
 
     def test_no_project(self):
-        with self.assertRaises(dbt.config.DbtProjectError) as exc:
+        with self.assertRaises(dbt.exceptions.DbtProjectError) as exc:
             dbt.config.Project.from_project_root(self.project_dir)
 
         self.assertIn('no dbt_project.yml', str(exc.exception))
@@ -690,7 +691,7 @@ class TestProjectFile(BaseFileTest):
 
     def test_with_invalid_package(self):
         self.write_packages({'invalid': ['not a package of any kind']})
-        with self.assertRaises(dbt.config.DbtProjectError) as exc:
+        with self.assertRaises(dbt.exceptions.DbtProjectError) as exc:
             dbt.config.Project.from_project_root(self.project_dir)
 
 
@@ -744,7 +745,7 @@ class TestRuntimeConfig(BaseConfigTest):
         profile = self.get_profile()
         # invalid - must be boolean
         profile.use_colors = None
-        with self.assertRaises(dbt.config.DbtProjectError):
+        with self.assertRaises(dbt.exceptions.DbtProjectError):
             dbt.config.RuntimeConfig.from_parts(project, profile, {})
